src/app/api/upload/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==2.0.0
python-multipart==0.0.6
python-dotenv==1.0.0
psycopg2-binary==2.9.7
